requirements.txt
Flask==3.1.2
Flask-CORS==6.0.1
SQLAlchemy==2.0.44
psutil==7.1.0
python-nmap==0.7.1
APScheduler==3.11.0
gunicorn==22.0.0
scapy==2.5.0
python-dotenv==1.0.0
werkzeug==3.1.3
requests==2.31.0